requirements.txt
pandas
pyarrow
python-dotenv
google-api-python-client
